+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import checkr ggplot2
+#' @import checkr ggplot2 chk
 #' @importFrom stats setNames
 #' @importFrom graphics plot
 NULL